requirements.txt
pbr!=2.1.0
PyYAML==5.1.2
tornado==6.0.2
juju==0.11.5
python-libmaas==0.6.4
